+++ examples/quickstart/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @machinen/example-quickstart
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @machinen/runtime@0.3.1
+
 ## 0.0.5
 
 ### Patch Changes